backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Kitahack - AI Learning StoryBook Backend with Google Cloud"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn[standard]>=0.41.0",
    "google-cloud-storage>=2.10.0",
    "google-cloud-aiplatform>=1.40.0",
    "google-auth>=2.25.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
]
